+++ packages/layout/src/components/Layout/common/Toggle/ToggleMenuMobile.tsx
@@ -1,6 +1,6 @@
 import { useMediaQuery } from "@dzangolab/react-ui";
 
-import { useLayoutContext } from "../../Context";
+import { useLayoutContext } from "../../../../context/LayoutProvider";
 
 export const ToggleMenuMobile = () => {
   const {
