+++ packages/layout/src/components/Layout/common/Toggle/ToggleSidebar.tsx
@@ -1,4 +1,4 @@
-import { useLayoutContext } from "../../Context";
+import { useLayoutContext } from "../../../../context/LayoutProvider";
 
 export const ToggleSidebar = () => {
   const { menuDesktopOpen, setMenuDesktopOpen } = useLayoutContext();
